revert changes in untested endpoints
